+++ tests/test_shell_completion.py
@@ -161,20 +161,6 @@ def test_option_custom():
     assert _get_words(cli, ["a", "b"], "c") == ["C"]
 
 
-def test_autocompletion_deprecated():
-    # old function takes args and not param, returns all values, can mix
-    # strings and tuples
-    def custom(ctx, args, incomplete):
-        assert isinstance(args, list)
-        return [("art", "x"), "bat", "cat"]
-
-    with pytest.deprecated_call():
-        cli = Command("cli", params=[Argument(["x"], autocompletion=custom)])
-
-    assert _get_words(cli, [], "") == ["art", "bat", "cat"]
-    assert _get_words(cli, [], "c") == ["cat"]
-
-
 def test_option_multiple():
     cli = Command(
         "type",